proxy: Remove socket if file exists
The clc_snippet creates the file
/var/konnectivity-server/konnectivity-server.socket. So, the first time
the proxy is deployed, the file exists and fails to start. We need to
manually remove it for it to start.

Using this flag, the file is removed and just works. This way it would
also work if, for some reason, that file isn't cleaned up in some case

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>