konnectivity server: Allow to deploy on non-HA setups
Before this change, it will try to create the new pod and only after
that kill the old pod.

However, on setups with only one controller that doesn't work:
 * The yaml has a nodeSelector to run only on the controller
 * It uses hostNetwork.

Therefore, the new pod is never scheduled to the node (as it is using
the same ports we want to use, as it uses hostNetwork) and the deploy is
stalled in pending.

Allowing to kill one pod first, as this patch does, allos the deployment
to work in single master setups.